第一次整合

+++ bakery/src/com/product/controller/ProductInsertServlet.java
@@ -71,13 +71,13 @@ public class ProductInsertServlet extends HttpServlet {
 			
 			fi.close();
 
-			double discount = 0;
+			float discount = 0;
 			
 			String temp3 = request.getParameter("discount");
 			if (temp3 == null || (temp3.trim()).length() == 0) {
 				errors.put("discount", "請輸入折扣");
 			} else {
-				discount = Double.parseDouble(temp3);
+				discount = Float.parseFloat(temp3);
 			}
 
 			String temp4 = request.getParameter("productDate");
